+++ D3D11Manager/D3D11Manager/HullShader.cpp
@@ -36,5 +36,5 @@ void CHullShader::ResetSamplerState(ID3D11DeviceContext* deviceContext) const no
 {
 	vector<ID3D11SamplerState*> resetSamplerState;
 	resetSamplerState.resize(m_samplerStateCount, nullptr);
-	deviceContext->HSSetSamplers(0, static_cast<UINT>(m_samplerStateCount), resetSamplerState.data());
+	deviceContext->HSSetSamplers(0, m_samplerStateCount, resetSamplerState.data());
 }
