+++ D3D11Manager/D3D11Manager/PixelShader.cpp
@@ -36,5 +36,5 @@ void CPixelShader::ResetSamplerState(ID3D11DeviceContext* deviceContext) const n
 {
 	vector<ID3D11SamplerState*> resetSamplerState;
 	resetSamplerState.resize(m_samplerStateCount, nullptr);
-	deviceContext->PSSetSamplers(0, static_cast<UINT>(m_samplerStateCount), resetSamplerState.data());
+	deviceContext->PSSetSamplers(0, m_samplerStateCount, resetSamplerState.data());
 }
